Add delay information to CompletableFuture
